requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==2.0.3
llama-cpp-python==0.2.11